Refactor component test namespaces.
